+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - New component: `Doggo.menu_button/1`.
 - New component: `Doggo.menu_group/1`.
 - New component: `Doggo.menu_item/1`.
+- New component: `Doggo.menu_item_checkbox/1`.
 - New component: `Doggo.radio_group/1`.
 - New component: `Doggo.split_pane/1`.
 - New component: `Doggo.tabs/1`.